Make one test pass.
This is getting a bit hairy.  Of course, encapsulating the complexity is
what this package is for... however, since it now seems I need a
go-routine per out and in/out socket to abstract this properly, it may
make more sense to provide this functionality as a simpler extension of
zmq.Socket.