Panic in NewContext and NewSocket.
Modify NewSocket and NewContext to panic instead of returning errors.
The rationale is that an error will occur only if either there is not
enough memory available, some other built-in limit (e.g. on maximum
number of sockets per context) is reached, or the provided arguments are
invalid.

This commit updates the documented example code to suit this change,
also correcting it to work at all (as it did not previously).